MLBEDSW-8569: MLCE: Reported number of CPU ops are wrong
 - A Pack op is implemented by several AvgPool ops. Depending
on number of CPU ops and graph topology this could result in that
the AvgPool ops ended up in different nodes. One of these node
had the Pack output referenced to it but the other node did not.
As a result the full graph was not traversed when calculating CPU
ops.
 - The compiled network works as intended but the number of
reported CPU was wrong.
 - Added new method that extracts the ops using the passes in
the sub graphs which fix the problem.

Change-Id: Ie88ebd4669783559258ae763737a4c7f86c905f8
Signed-off-by: Johan Alfven <johan.alfven@arm.com>