MLBEDSW-8875: MLCE: Update criteria when to move SplitSpliceRead to consumer
 - When possible, a read slice from a split or stride is moved to
the following op. The problem in this case was that the following
op was a Maxpool op (from Softmax). The Maxpool op is using a
different input shape compared to the original Softmax op, and
this input shape was then changed when the read slice was applied
to the Maxpool op.
 - The result is a faulty Maxpool op with an output diff.
 - The fix is to prevent moving the slice read when the consumer
input shape differs from the Split/Stride ofm shape

Change-Id: I649d89c38645fa51c20c3602954e2b8af9372076
Signed-off-by: Johan Alfven <johan.alfven@arm.com>