MLBEDSW-8885: MLCE: Fix assert in verify_subgraph_health
 - Assert triggered due to that the tensor consumer list did
not contain expected operators.
 - The problem happened because a concat op was split into two
avgpool ops and these two ops run in separate subgraphs with
a cpu node in between. Since the avgpool ops share the same
output tensor this caused some corruption to the tensor consumer
list when the last subgraph was traversed.
 - The fix is to ignore ops that do not belong in the subgraph's
set of operators (the pass list) when updating the consumers.

Change-Id: I4d94b54c77001f6447bec31ec62daeebc9b104f9
Signed-off-by: Johan Alfven <johan.alfven@arm.com>