MLBEDSW-8986 Avoid unnecessary weight encoding
Identify if sparsity and/or fast encoder can be discarded
before actual encode.
Collect statistics on weights during default encode, so size
of FWD stream can be calculated, and only do the FWD encode if
the FWD has been selected.
Don't do the encodes in measure-only mode, it has bugs, and the
work needs to be redone for the actual encode anyway.

Change-Id: I1dabbc87b1e268dcecb4fbffe9f67a18563eec08
Signed-off-by: Björn Davidsson <bjoern.davidsson@arm.com>