MLBEDSW-9623 Add support for TOSA VARIABLE
Added support for TOSA VARIABLE tensors by extending the live range
for variables to the duration of the entire inference.
Made sure variable tensors are added as an input to the custom
EthosU op in order to signal the existense of the variable to the
runtime. Also added a new tensor type for variables to the raw_writer.

Change-Id: I16b8cf573271e1ecc71057ae6ce8d6cdb3264c26
Signed-off-by: Fredrik Svedberg <fredrik.svedberg@arm.com>