MLBEDSW-11395: Use conn shape when calculating rolling buffer shape
If we have a cascade of ops where the intermediate IFM tensor storage
shape is not matching the IFM tensor connection shape in the width
dimension (for example due to reshapes), this assert will trigger. It's
not possible to construct a such network in TOSA or TFLite, but it's
possible to end up in this situation after optimizations.

Example:

Ta -> OP1 -> Tb -> OP2 -> Tc

Ta storage shape: [1, 8, 385, 1]
OP1 IFM conn shape: [1, 11, 7, 40]
OP1 OFM conn shape: [1, 11, 7, 40]
Tb storage shape: [1, 8, 385, 1]
OP2 IFM conn shape: [1, 11, 7, 40]
OP2 OFM conn shape: [1, 11, 7, 40]
Tc storage shape: [1, 8, 385, 1]

In this example we will trigger the issue when OP1 and OP2 are cascaded
and when we reach OP2 in the ApplySchedule function.

Signed-off-by: Johan Gunnarsson <johan.gunnarsson@arm.com>
Change-Id: If4cb30cea85a809950be5a7e8197617477f932c3