MLBEDSW-9780: Add placeholder tensors to unconnected ops
The TFLite CallOnce and TFLite AssignVariable has no OFM and will
therefore not be reachable when traversing the graph starting from
the graph output tensors, but we still need to keep them in the
graph so we can pass them to CPU. This change discovers such
unconnected ops when loading the graph, and adds a placeholder
output tensors to them and stores the placeholder tensors in both
graph outputs and in a separate list. Later when writing out this
graph, remove these placeholder tensors again.

Do the same for ops without input tensor.

Signed-off-by: Johan Gunnarsson <johan.gunnarsson@arm.com>
Change-Id: I68c963c8b9c447527eb16ef415b85ac0253f9ff6