MLBEDSW-10228: Remove ConstraintSliceRanges
This check is not correct. The begin/end/strides vectors are not
guaranteed have the same length as number of dimensions in IFM.
Also, it's not possible to implement a proper check of
begin/end/strides since these tensors are not guaranteed to be
constant or known at compile time.

Signed-off-by: Johan Gunnarsson <johan.gunnarsson@arm.com>
Change-Id: Id4326e05584c81929bc5fd9aafcc584319e10d0b