MLBEDSW-10158: Don't allocate fast memory for cross-CPU tensors
A cross-CPU tensor is a tensor with a live range that covers two
command streams, potentially with a CPU op in between them. In a
multi-threaded setup, the fast storage is not guaranteed to be
left unchanged between the execution of the command streams.

This patch can have a negative performance impact for networks
with multiple command streams.

Signed-off-by: Johan Gunnarsson <johan.gunnarsson@arm.com>
Change-Id: I1cc8f51bf5c1b01e212dedc86c9cd2a415648ec0