MLBEDSW-10285: MLCE: Output diff caused by wrong ifm box
 - A slice op followed by a conv2d with stride 2 caused an output diff
 - The slice read is moved to the consumer (conv2d) but the problem in
this case was that the ifm box calculation was not correct when having
a stride greater than one
 - The issue is solved by backporting various fixes from Regor that is
making sure ifm and ofm box have correct offsets and sizes
 - Also fixed a hidden problem that read_shape in rewrite_split_ops was
calculated erroneously since start and end offset can be less than
rank 4 but ifm shape is always 4. That gave a corrupt read_shape.
However, read_shape height was not used before this commit so corrupt
value was not used and did not cause any problems

Change-Id: Ib71c13cfecf77b2cdc2b5aaf437938577c433bb5
Signed-off-by: Johan Alfvén <johan.alfven@arm.com>