MLBEDSW-10249 Fixed scaling issue with out-of-range shifts
When the shift is above 63 it cannot be encoded and would cause an
assert. However, a shift larger than 63 would lead to all the bits being
shifted away and result in a 0 anyway so the same behaviour can be
achieved by setting the scale to 0.

Also addressed two minor integer type issues in the python graph
optimiser.

Change-Id: I5da2b4170f95c4e1cf161c93a932c92ad9c242ea
Signed-off-by: Jacob Bohlin <jacob.bohlin@arm.com>