MLBEDSW-10619: Fix unbuffered weights regression
This fixes unbuffered weights for multicore configurations of
Ethos-U85. Unbuffered weights are never depth sliced so we
should never add range.offset. On Ethos-U55, range.offset is
always 0 for unbuffered weights so it didn't matter. On Ethos-U85
multicore configurations, range.offset is not always 0 even for
unbuffered weights so that caused the address to be offset too
much.

This commit partially reverts commit ae708a2.

Signed-off-by: Johan Gunnarsson <johan.gunnarsson@arm.com>
Change-Id: I212626773512aeeb8709be71a5627b29428dec58