MLBEDSW-10776 Revert graph traversal change
The graph traversal was modified in MLBEDSW-8926 to traverse tensor
writers left-to-right instead of right-to-left. This is required to
ensure correct execution order for LSTM.

This change reverts the graph traversal back to right-to-left in the
general case and left-to-right will only be used on graphs which contain
persistent tensors, in order to target LSTM operators.

Change-Id: Ibe13af0cf952450cff253ff2e44ee8b96068583a
Signed-off-by: Jacob Bohlin <jacob.bohlin@arm.com>