MLBEDSW-11132: Fix Depthwise weight layout issue when rewritten
A GraphIR optimisation that rewrites Depthwise into Conv2D (when
a multiplier is present) raised a corner-case issue in the
depthwise weight refactor causing some verification failures.

Rather than permuting the weight memory, the rewrite set a
non-standard axis layout that enabled the depthwise weights to
be transparently viewed as convolution weights via late striding
adjustment.

This commit allows the weight tensors to be specifically restrided
by rewrites earlier in the pipeline to be compatible with the depthwise
refactor.

 - Also applied patch to remove dynamic loop list missed in rework patch.

Signed-off-by: Philip Hall <philip.hall@arm.com>
Change-Id: Id0cf18f30e41e3430aceb78975f339dd1228cebb