MLBEDSW-10894: Add a perf_debug_conn table to debug database
The table describes connectivity between ops in Scheduler IR. Each
row describes one input of an op. It contains 3 fields:

* id: The scheduler IR operation ID.
* input_op_id: The scheduler IR operation ID of a node producing an
  input.
* input_index: The IFM index that is consuming this input.

Two special values are used to describe graph inputs and graph
outputs. "-2" is used in this "id" field for graph outputs. "-1" is
used in the "input_op_id" field for graph inputs.

A network that looks like this...

[T1] -> (A) -> [T2] -> (B) -> [T3]

...where T1, T2, T3 are tensors, T1 is graph input, T3 is graph
output, A (ID 0), B (ID 1) are operations, will have a
perf_debug_conn table that looks like this:

  id, input_op_id, input_index
   0,          -1,           0
   1,           0,           0
  -2,           1,           0

Signed-off-by: Johan Gunnarsson <johan.gunnarsson@arm.com>
Change-Id: Ia8cd8cad35ecf2d3e58987c651856a7d6ea8401b