MLBEDSW-10786: Handle multiple constant 1-element IFMs
Only one IFM can be used as scalar. If an op has multiple IFMs that
could be used as scalars, pick IFM0 and keep the other IFMs as tensors.
This decision whether an IFM should be scalar or not happens in RCS
generator. To be able to fall back to using read only 1-element IFMs
as tensors we need to put all tensors in read only memory, even if they
could potentially be ignored. This will waste small amounts of read only
memory.

Signed-off-by: Johan Gunnarsson <johan.gunnarsson@arm.com>
Change-Id: I9adf7bce615bf705eaf2686fd9640b27130fab17