MLBEDSW-10927: Fix identity scaling for pooling mode SUM
Right now, for avg_pool with global scaling when pooling mode sum is required, an implicit scale is set to implement averaging. When we don't want this implicit scaling, we clear the scales to signal this (e.g. in decomposition) and this patch makes sure we check explicitly for that in the rcs gen.

Change-Id: I8662de9315542906508dd8f97246f6b80c19eb46
Signed-off-by: William Isaksson <william.isaksson@arm.com>