MLBEDSW-11034: Set correct OFM shape for LUTs
We want LUTs to have same IFM and OFM shape in case the LUT ends up
not being fused later on. If it's not fused, it will be converted to
a pool op, which needs valid IFM and OFM shapes otherwise the IFM
area may be incorrect. We can not use OFM's storage shape, since it
can be an unexpected shape due to reshape ops.

Signed-off-by: Johan Gunnarsson <johan.gunnarsson@arm.com>
Change-Id: Ib924f397c3cba458d0968202ccb88519f9f8520d