MLBEDSW-11114 Decompose non-const convs into ops with OFM depth > 1
Flattens the shape of the non-constant weight tensor from:
[OC, KH, KW, IC] to [1, 1, OC*KH*KW, IC] and sets the stride to KH*KW in
order to step through the kernels correctly.
Change the opcode of the decomposed ops to MatMul instead of Conv as
this is more accurately what it's doing and leads to correct box
calculations. MatMul still maps to NPU_OP_CONV in the command stream.

Change-Id: Iacf377024772eacde4c7a3617c274808dfb9a995
Signed-off-by: Jacob Bohlin <jacob.bohlin@arm.com>