MLBEDSW-11193: Fix constprop when whole graph is constant
If all input to ta graph is constant, the graph should be replaced
by a single memory copy that copies a constant tensor to the graph
outout. Previously every op was removed and the graph output was
converted to a constant tensor. Empty graphs are problematic for us
since there is nothing left that writes to graph output.

Signed-off-by: Johan Gunnarsson <johan.gunnarsson@arm.com>
Change-Id: Ifdcd66ed14e772f6d8b37e3e3b110dede8222277