MLBEDSW-11171 Fix cell state rescaling issue for LSTM
At the end of every LSTM step there is an int16 tanh applied to the current
cell state. The tanh expects a range between [-1, 1] but the cell state
is not required to be quantized to the same range. This means a rescale
op has to be inserted before the tanh.

Change-Id: Ibbfaca9b5a1532f4af90d7aaadb19c8c7d909ff0
Signed-off-by: Jacob Bohlin <jacob.bohlin@arm.com>