MLBEDSW-11396: Connect more inputs to IF/WHILE through a memcopy
Connect all inputs to a control flow operator that are either
constant or graph input/output through a memory copy. This to make
sure that all input tensors to a IF/WHILE in a regular feature map
memory area. This will simplify the memory allocation because it then
only have to care about addresses and not memory area.

Signed-off-by: Johan Gunnarsson <johan.gunnarsson@arm.com>
Change-Id: I150dee1ff791363572913ec689e2c053337bf818